(fix) Using Google Maps Geolocation API for Zip Code validation for improved error handling, improved file naming practices in /api folder
