+++ src/app/components/Geolocation/Geolocation.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { addZipCode } from '../../../store/actions/zipCodes';
-import { geolocation } from '../../../api/googleGeolocation';
+import { geolocation } from '../../../api/googleMaps';
 
 const Geolocation: React.FC = () => {
   const dispatch = useDispatch();
