+++ src/helpers/locationValidation.ts
@@ -1,4 +1,4 @@
-import { fetchZipCodeValidation } from '../api/openWeatherMap';
+import { fetchZipCodeValidation } from '../api/googleMaps';
 const regexp = /^[0-9]{5}(?:-[0-9]{4})?$/;
 
 export async function locationValidation(
@@ -14,7 +14,8 @@ export async function locationValidation(
   } else if (!regexp.test(formInput)) {
     return 'invalid';
   } else {
-    if (await fetchZipCodeValidation(formInput)) {
+    const validation = await fetchZipCodeValidation(formInput);
+    if (validation) {
       return 'valid';
     } else {
       return 'invalid';
